CCM NG: Refactored all (hopefully) Repositories to return an Optional instead of a result or null for the methods returning a single entity as result.
git-svn-id: https://svn.libreccm.org/ccm/ccm_ng@4549 8810af33-2d31-482b-a856-94f89814c4df

Former-commit-id: 285cc3b6c59e312892b0f6372f0a63ff8a648915

+++ ccm-core/src/main/java/org/libreccm/categorization/DomainManager.java
@@ -24,6 +24,7 @@ import org.libreccm.web.ApplicationRepository;
 import org.libreccm.web.CcmApplication;
 
 import java.util.List;
+import java.util.Optional;
 
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
@@ -128,26 +129,45 @@ public class DomainManager {
     @Transactional(Transactional.TxType.REQUIRED)
     public void removeDomainOwner(final CcmApplication application,
                                   final Domain domain) {
-        final CcmApplication owner = applicationRepo.findById(
+        if (application == null) {
+            throw new IllegalArgumentException("Can't remove owner null.");
+        }
+        if (domain == null) {
+            throw new IllegalArgumentException(
+                "Can't remove the owner from domain null.");
+        }
+
+        final Optional<CcmApplication> owner = applicationRepo.findById(
             application.getObjectId(), "CcmApplication.withDomains");
-        final Domain ownedDomain = domainRepo.findById(
+        final Optional<Domain> ownedDomain = domainRepo.findById(
             domain.getObjectId(), "Domain.withOwners");
+
+        if (!owner.isPresent()) {
+            throw new IllegalArgumentException(String.format(
+                "The provided owner %s does not exist in the database.",
+                application.toString()));
+        }
+        if (!ownedDomain.isPresent()) {
+            throw new IllegalArgumentException(String.format(
+                "The provided domain %s does not exist in the database.",
+                domain.toString()));
+        }
 
         final TypedQuery<DomainOwnership> query = entityManager
             .createNamedQuery("DomainOwnership.findByOwnerAndDomain",
                               DomainOwnership.class);
-        query.setParameter("owner", owner);
-        query.setParameter("domain", ownedDomain);
+        query.setParameter("owner", owner.get());
+        query.setParameter("domain", ownedDomain.get());
 
         final List<DomainOwnership> result = query.getResultList();
 
         if (result != null) {
             result.forEach(o -> {
-                ownedDomain.removeOwner(o);
-                owner.removeDomain(o);
+                ownedDomain.get().removeOwner(o);
+                owner.get().removeDomain(o);
                 entityManager.remove(o);
-                domainRepo.save(ownedDomain);
-                applicationRepo.save(owner);
+                domainRepo.save(ownedDomain.get());
+                applicationRepo.save(owner.get());
             });
         }
     }
